Refactor CampaignSelect
