--- src/pages/CampaignSelect/CampaignModal/CampaignModal.scss
@@ -1,6 +1,0 @@
-@import "../../../common/variables";
-
-.CampaignModalButton {
-    margin-top: $modal-spacing;
-    align-self: flex-end;
-}
