--- src/pages/CampaignSelect/CampaignSelect.scss
@@ -1,72 +1,0 @@
-@import "../../common/theme";
-@import "../../common/variables";
-
-// gta PAGE
-.CampaignSelect {
-  display: grid;
-  padding-top: 20px;
-  grid-template-rows: 20px 60% 20%;
-  grid-template-columns: 20% 60% 20%;
-  grid-template-areas:
-    ". . ."
-    ". content ."
-    ". . .";
-
-  @media (max-width: $landscape-breakpoint) {
-    grid-template-rows: 20px 60% 20%;
-    grid-template-columns: 10% 80% 10%;
-    grid-template-areas:
-      ". . ."
-      ". content ."
-      ". . .";
-  }
-  @media (max-width: $tablet-breakpoint) {
-    grid-template-rows: 20px 60% 20%;
-    grid-template-columns: 100%;
-    grid-template-areas:
-      "."
-      "content"
-      ".";
-  }
-  @include themed() {
-    background-color: t($background);
-    color: t($background);
-  }
-}
-
-.CampaignSelectContent {
-  grid-area: content;
-  @include themed() {
-    color: t($on-background);
-  }
-}
-
-.CampaignSelectHeader {
-  margin-bottom: 30px;
-}
-
-.CampaignSelectNew {
-  margin-bottom: 10px;
-  padding: 0.2em 1em;
-  @include themed() {
-    &:hover {
-      cursor: pointer;
-      color: desaturate(t($on-default), 70%);
-    }
-    background-color: t($default);
-  }
-}
-
-.CampaignSelectListItem {
-  // override default .Text styles
-  padding: 1em 0 !important;
-}
-
-.CampaignSelectListItemIcon {
-  padding: 0 1em;
-}
-
-.CampaignModal {
-  display: flex;
-  flex-direction: column;
-}
